Move help service back to main folder
